backend/requirements.txt
flask==3.0.0
filelock>=3.13
flask-cors==4.0.0
flask-limiter>=3.5.0
redis>=5.0
pydantic>=2.0
kiteconnect==4.2.0
python-dotenv==1.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0
pandas>=2.2.0
beautifulsoup4>=4.12.2
requests>=2.31.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
APScheduler>=3.10.4
pytz>=2024.1
anthropic>=0.40.0
cryptography>=42.0.0
yfinance>=0.2.0
gunicorn>=22.0.0
sentry-sdk[flask]>=2.0.0

# Dev / CI
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.4.0
